+++ cmd/moombox/main.go
@@ -43,7 +43,7 @@ import (
 )
 
 var (
-	version = "2.1.8"
+	version = "2.1.9"
 	commit  = ""
 )